@@ -35,92 +36,61 @@
   "author": "Brandon Keepers",
   "license": "ISC",
   "dependencies": {
-    "@octokit/core": "^3.2.4",
-    "@octokit/plugin-enterprise-compatibility": "^1.2.8",
-    "@octokit/plugin-paginate-rest": "^2.6.2",
-    "@octokit/plugin-rest-endpoint-methods": "^5.0.1",
-    "@octokit/plugin-retry": "^3.0.6",
-    "@octokit/plugin-throttling": "^3.3.4",
-    "@octokit/types": "^8.0.0",
-    "@octokit/webhooks": "^9.8.4",
-    "@probot/get-private-key": "^1.1.0",
-    "@probot/octokit-plugin-config": "^1.0.0",
-    "@probot/pino": "^2.2.0",
-    "@types/express": "^4.17.9",
-    "@types/ioredis": "^4.27.1",
-    "@types/pino": "^6.3.4",
-    "@types/pino-http": "^5.0.6",
-    "commander": "^6.2.0",
-    "deepmerge": "^4.2.2",
-    "deprecation": "^2.3.1",
-    "dotenv": "^8.2.0",
-    "eventsource": "^2.0.2",
-    "express": "^4.17.1",
-    "express-handlebars": "^6.0.3",
-    "ioredis": "^4.27.8",
-    "js-yaml": "^3.14.1",
-    "lru-cache": "^6.0.0",
-    "octokit-auth-probot": "^1.2.2",
-    "pino": "^6.7.0",
-    "pino-http": "^5.3.0",
+    "@octokit/core": "^5.0.2",
+    "@octokit/plugin-enterprise-compatibility": "^4.0.1",
+    "@octokit/plugin-paginate-rest": "^9.1.4",
+    "@octokit/plugin-rest-endpoint-methods": "^10.1.5",
+    "@octokit/plugin-retry": "^6.0.1",
+    "@octokit/plugin-throttling": "^8.1.3",
+    "@octokit/request": "^8.1.6",
+    "@octokit/types": "^12.3.0",
+    "@octokit/webhooks": "^12.0.10",
+    "@probot/get-private-key": "^1.1.2",
+    "@probot/octokit-plugin-config": "^2.0.1",
+    "@probot/pino": "^2.5.0",
+    "@types/express": "^4.17.21",
+    "bottleneck": "^2.19.5",
+    "commander": "^12.0.0",
+    "deepmerge": "^4.3.1",
+    "dotenv": "^16.3.1",
+    "express": "^4.21.0",
+    "ioredis": "^5.3.2",
+    "js-yaml": "^4.1.0",
+    "lru-cache": "npm:@wolfy1339/lru-cache@^11.0.2-patch.1",
+    "octokit-auth-probot": "^2.0.1",
+    "pino": "^9.0.0",
+    "pino-http": "^10.0.0",
     "pkg-conf": "^3.1.0",
-    "resolve": "^1.19.0",
-    "semver": "^7.3.4",
-    "update-dotenv": "^1.1.1",
-    "uuid": "^8.3.2"
+    "update-dotenv": "^1.1.1"
   },
   "devDependencies": {
-    "@octokit/webhooks-examples": "^4.0.0",
-    "@octokit/webhooks-methods": "^3.0.0",
-    "@octokit/webhooks-types": "^4.0.0",
-    "@tsconfig/node10": "^1.0.7",
-    "@types/eventsource": "^1.1.5",
-    "@types/jest": "^26.0.18",
-    "@types/js-yaml": "^3.12.5",
-    "@types/jsonwebtoken": "^8.5.0",
-    "@types/node": "^16.0.0",
-    "@types/readable-stream": "^2.3.9",
-    "@types/resolve": "^1.17.1",
-    "@types/semver": "^7.3.4",
-    "@types/supertest": "^2.0.10",
-    "@types/uuid": "^8.3.0",
-    "body-parser": "^1.19.0",
-    "bottleneck": "^2.19.5",
+    "@octokit/tsconfig": "^3.0.0",
+    "@octokit/webhooks-examples": "^7.3.1",
+    "@octokit/webhooks-methods": "^4.0.0",
+    "@octokit/webhooks-types": "^7.3.1",
+    "@types/js-yaml": "^4.0.9",
+    "@types/node": "^22.0.0",
+    "@types/supertest": "^6.0.0",
+    "@vitest/coverage-v8": "^3.0.0",
+    "@vitest/ui": "^3.0.0",
     "connect-sse": "^1.2.0",
     "execa": "^5.0.0",
+    "fetch-mock": "^11.0.0",
     "get-port": "^5.1.1",
-    "got": "^11.8.0",
-    "jest": "^26.6.3",
-    "nock": "^13.0.5",
-    "prettier": "^2.2.1",
-    "rimraf": "^3.0.2",
-    "semantic-release": "^19.0.3",
+    "prettier": "^3.0.3",
+    "rimraf": "^6.0.0",
+    "semantic-release": "^24.0.0",
     "semantic-release-plugin-update-version-in-files": "^1.1.0",
-    "smee-client": "^1.2.2",
-    "supertest": "^6.0.1",
-    "ts-jest": "^26.4.4",
-    "tsd": "^0.23.0",
-    "typedoc": "^0.22.10",
-    "typescript": "^4.1.2"
+    "smee-client": "^2.0.0",
+    "sonic-boom": "^4.0.0",
+    "supertest": "^7.0.0",
+    "tsd": "^0.31.0",
+    "typedoc": "^0.27.0",
+    "typescript": "^5.2.2",
+    "vitest": "^3.0.0"
   },
   "engines": {
-    "node": ">=10.21"
-  },
-  "jest": {
-    "testEnvironment": "node",
-    "coveragePathIgnorePatterns": [
-      "<rootDir>/lib/"
-    ],
-    "moduleFileExtensions": [
-      "ts",
-      "js",
-      "json",
-      "node"
-    ],
-    "preset": "ts-jest"
-  },
-  "tsd": {
-    "directory": "test/types"
+    "node": ">=18"
   },
   "release": {
     "plugins": [
