@@ -1,8 +1,7 @@
 import type { Logger } from "pino";
-
-import { getAuthenticatedOctokit } from "./octokit/get-authenticated-octokit";
-import { ProbotOctokit } from "./octokit/probot-octokit";
-import { State } from "./types";
+import { getAuthenticatedOctokit } from "./octokit/get-authenticated-octokit.js";
+import { ProbotOctokit } from "./octokit/probot-octokit.js";
+import type { State } from "./types.js";
 
 /**
  * Authenticate and get a GitHub client that can be used to make API calls.
@@ -14,12 +13,13 @@ import { State } from "./types";
  * to wait for the magic to happen.
  *
  * ```js
- *  module.exports = (app) => {
+ *  export default (app) => {
  *    app.on('issues.opened', async context => {
  *      const octokit = await app.auth();
  *    });
  *  };
  * ```
+ * @param state - Probot application instance state, which is used to persist
  *
  * @param id - ID of the installation, which can be extracted from
  * `context.payload.installation.id`. If called without this parameter, the
@@ -32,10 +32,7 @@ import { State } from "./types";
 export async function auth(
   state: State,
   installationId?: number,
-  log?: Logger
+  log?: Logger,
 ): Promise<InstanceType<typeof ProbotOctokit>> {
-  return getAuthenticatedOctokit(
-    Object.assign({}, state, log ? { log } : null),
-    installationId
-  );
+  return getAuthenticatedOctokit(Object.assign({}, state), installationId, log);
 }
