@@ -1,14 +1,12 @@
-import path from "path";
-
-import { EmitterWebhookEvent as WebhookEvent } from "@octokit/webhooks";
+import path from "node:path";
 import merge from "deepmerge";
 
+import type {
+  EmitterWebhookEvent as WebhookEvent,
+  EmitterWebhookEventName as WebhookEvents,
+} from "@octokit/webhooks";
 import type { Logger } from "pino";
-
-import { ProbotOctokit } from "./octokit/probot-octokit";
-import { aliasLog } from "./helpers/alias-log";
-import { DeprecatedLogger } from "./types";
-import { EmitterWebhookEventName as WebhookEvents } from "@octokit/webhooks/dist-types/types";
+import type { ProbotOctokit } from "./octokit/probot-octokit.js";
 
 export type MergeOptions = merge.Options;
 
@@ -31,12 +29,12 @@ type RepoIssueNumberType<T extends WebhookEvents> =
   WebhookEvent<T>["payload"] extends { issue: { number: number } }
     ? number
     : never | WebhookEvent<T>["payload"] extends {
-        pull_request: { number: number };
-      }
-    ? number
-    : never | WebhookEvent<T>["payload"] extends { number: number }
-    ? number
-    : never;
+          pull_request: { number: number };
+        }
+      ? number
+      : never | WebhookEvent<T>["payload"] extends { number: number }
+        ? number
+        : never;
 
 /** Context.repo return type */
 type RepoResultType<E extends WebhookEvents> = {
@@ -49,7 +47,7 @@ type RepoResultType<E extends WebhookEvents> = {
  * helpers for extracting information can be passed to GitHub API calls.
  *
  *  ```js
- *  module.exports = app => {
+ *  export default app => {
  *    app.on('push', context => {
  *      context.log.info('Code was pushed to the repo, what should we do with it?');
  *    });
@@ -60,25 +58,33 @@ type RepoResultType<E extends WebhookEvents> = {
  * @property {payload} payload - The webhook event payload
  * @property {log} log - A pino instance
  */
-export class Context<E extends WebhookEvents = WebhookEvents> {
+export class Context<Event extends WebhookEvents = WebhookEvents> {
   public name: WebhookEvents;
   public id: string;
-  public payload: WebhookEvent<E>["payload"];
+  public payload: {
+    [K in Event]: K extends WebhookEvents ? WebhookEvent<K> : never;
+  }[Event]["payload"];
 
-  public octokit: InstanceType<typeof ProbotOctokit>;
-  public log: DeprecatedLogger;
+  public octokit: ProbotOctokit;
+  public log: Logger;
 
-  constructor(
-    event: WebhookEvent<E>,
-    octokit: InstanceType<typeof ProbotOctokit>,
-    log: Logger
-  ) {
+  constructor(event: WebhookEvent<Event>, octokit: ProbotOctokit, log: Logger) {
     this.name = event.name;
     this.id = event.id;
     this.payload = event.payload;
 
     this.octokit = octokit;
-    this.log = aliasLog(log);
+    this.log = log;
+
+    // set `x-github-delivery` header on all requests sent in response to the current
+    // event. This allows GitHub Support to correlate the request with the event.
+    // This is not documented and not considered public API, the header may change.
+    // Once we document this as best practice on https://docs.github.com/en/rest/guides/best-practices-for-integrators
+    // we will make it official
+    /* istanbul ignore next */
+    octokit.hook.before("request", (options) => {
+      options.headers["x-github-delivery"] = event.id;
+    });
   }
 
   /**
@@ -93,13 +99,13 @@ export class Context<E extends WebhookEvents = WebhookEvents> {
    * @param object - Params to be merged with the repo params.
    *
    */
-  public repo<T>(object?: T): RepoResultType<E> & T {
-    // @ts-ignore `repository` is not always present in this.payload
+  public repo<T>(object?: T): RepoResultType<Event> & T {
+    // @ts-expect-error `repository` is not always present in this.payload
     const repo = this.payload.repository;
 
     if (!repo) {
       throw new Error(
-        "context.repo() is not supported for this webhook event."
+        "context.repo() is not supported for this webhook event.",
       );
     }
 
@@ -108,7 +114,7 @@ export class Context<E extends WebhookEvents = WebhookEvents> {
         owner: repo.owner.login,
         repo: repo.name,
       },
-      object
+      object,
     );
   }
 
@@ -125,16 +131,16 @@ export class Context<E extends WebhookEvents = WebhookEvents> {
    * @param object - Params to be merged with the issue params.
    */
   public issue<T>(
-    object?: T
-  ): RepoResultType<E> & { issue_number: RepoIssueNumberType<E> } & T {
+    object?: T,
+  ): RepoResultType<Event> & { issue_number: RepoIssueNumberType<Event> } & T {
     return Object.assign(
       {
         issue_number:
-          // @ts-ignore - this.payload may not have `issue` or `pull_request` keys
+          // @ts-expect-error - this.payload may not have `issue` or `pull_request` keys
           (this.payload.issue || this.payload.pull_request || this.payload)
             .number,
       },
-      this.repo(object)
+      this.repo(object),
     );
   }
 
@@ -151,15 +157,15 @@ export class Context<E extends WebhookEvents = WebhookEvents> {
    * @param object - Params to be merged with the pull request params.
    */
   public pullRequest<T>(
-    object?: T
-  ): RepoResultType<E> & { pull_number: RepoIssueNumberType<E> } & T {
+    object?: T,
+  ): RepoResultType<Event> & { pull_number: RepoIssueNumberType<Event> } & T {
     const payload = this.payload;
     return Object.assign(
       {
-        // @ts-ignore - this.payload may not have `issue` or `pull_request` keys
+        // @ts-expect-error - this.payload may not have `issue` or `pull_request` keys
         pull_number: (payload.issue || payload.pull_request || payload).number,
       },
-      this.repo(object)
+      this.repo(object),
     );
   }
 
@@ -225,25 +231,25 @@ export class Context<E extends WebhookEvents = WebhookEvents> {
   public async config<T>(
     fileName: string,
     defaultConfig?: T,
-    deepMergeOptions?: MergeOptions
+    deepMergeOptions?: MergeOptions,
   ): Promise<T | null> {
     const params = this.repo({
       path: path.posix.join(".github", fileName),
       defaults(configs: object[]) {
         const result = merge.all(
           [defaultConfig || {}, ...configs],
-          deepMergeOptions
+          deepMergeOptions,
         );
 
         return result;
       },
     });
 
-    // @ts-ignore
+    // @ts-expect-error
     const { config, files } = await this.octokit.config.get(params);
 
     // if no default config is set, and no config files are found, return null
-    if (!defaultConfig && !files.find((file: any) => file.config !== null)) {
+    if (!defaultConfig && !files.find((file) => file.config !== null)) {
       return null;
     }
 
