@@ -1,3 +1,3 @@
-import { run } from "../";
+import { run } from "../index.js";
 
 run(process.argv);
