@@ -1,65 +1,65 @@
-import program from "commander";
+import { program } from "commander";
 import { getPrivateKey } from "@probot/get-private-key";
-import { Options as PinoOptions } from "@probot/pino";
+import type { Options as PinoOptions } from "@probot/pino";
 
-import { Options } from "../types";
+import type { Options } from "../types.js";
 
 export function readCliOptions(
-  argv: string[]
+  argv: string[],
 ): Options & PinoOptions & { args: string[] } {
   program
     .usage("[options] <apps...>")
     .option(
       "-p, --port <n>",
       "Port to start the server on",
-      String(process.env.PORT || 3000)
+      String(process.env.PORT || 3000),
     )
     .option("-H --host <host>", "Host to start the server on", process.env.HOST)
     .option(
       "-W, --webhook-proxy <url>",
       "URL of the webhook proxy service.`",
-      process.env.WEBHOOK_PROXY_URL
+      process.env.WEBHOOK_PROXY_URL,
     )
     .option(
       "-w, --webhook-path <path>",
       "URL path which receives webhooks. Ex: `/webhook`",
-      process.env.WEBHOOK_PATH
+      process.env.WEBHOOK_PATH,
     )
     .option("-a, --app <id>", "ID of the GitHub App", process.env.APP_ID)
     .option(
       "-s, --secret <secret>",
       "Webhook secret of the GitHub App",
-      process.env.WEBHOOK_SECRET
+      process.env.WEBHOOK_SECRET,
     )
     .option(
       "-P, --private-key <file>",
       "Path to private key file (.pem) for the GitHub App",
-      process.env.PRIVATE_KEY_PATH
+      process.env.PRIVATE_KEY_PATH,
     )
     .option(
       "-L, --log-level <level>",
       'One of: "trace" | "debug" | "info" | "warn" | "error" | "fatal"',
-      process.env.LOG_LEVEL || "info"
+      process.env.LOG_LEVEL || "info",
     )
     .option(
       "--log-format <format>",
       'One of: "pretty", "json"',
-      process.env.LOG_FORMAT
+      process.env.LOG_FORMAT,
     )
     .option(
       "--log-level-in-string",
       "Set to log levels (trace, debug, info, ...) as words instead of numbers (10, 20, 30, ...)",
-      process.env.LOG_LEVEL_IN_STRING === "true"
+      process.env.LOG_LEVEL_IN_STRING === "true",
     )
     .option(
       "--sentry-dsn <dsn>",
       'Set to your Sentry DSN, e.g. "https://1234abcd@sentry.io/12345"',
-      process.env.SENTRY_DSN
+      process.env.SENTRY_DSN,
     )
     .option(
       "--redis-url <url>",
       'Set to a "redis://" url in order to enable cluster support for request throttling. Example: "redis://:secret@redis-123.redislabs.com:12345/0"',
-      process.env.REDIS_URL
+      process.env.REDIS_URL,
     )
     .option(
       "--base-url <url>",
@@ -68,7 +68,7 @@ export function readCliOptions(
         ? `${process.env.GHE_PROTOCOL || "https"}://${
             process.env.GHE_HOST
           }/api/v3`
-        : "https://api.github.com"
+        : "https://api.github.com",
     )
     .parse(argv);
 
@@ -77,12 +77,13 @@ export function readCliOptions(
     privateKey: privateKeyPath,
     redisUrl,
     ...options
-  } = program;
+  } = program.opts();
 
   return {
     privateKey: getPrivateKey({ filepath: privateKeyPath }) || undefined,
     appId,
     redisConfig: redisUrl,
+    args: program.args,
     ...options,
   };
 }
