@@ -1,16 +1,17 @@
-import { RequestListener } from "http";
-import { createNodeMiddleware as createWebbhooksMiddleware } from "@octokit/webhooks";
+import type { RequestListener } from "http";
+import { createNodeMiddleware as createWebhooksMiddleware } from "@octokit/webhooks";
 
-import { ApplicationFunction } from "./types";
-import { MiddlewareOptions } from "./types";
+import type { ApplicationFunction, MiddlewareOptions } from "./types.js";
+import { defaultWebhooksPath } from "./server/server.js";
+import { createProbot } from "./create-probot.js";
 
 export function createNodeMiddleware(
   appFn: ApplicationFunction,
-  { probot, webhooksPath }: MiddlewareOptions
+  { probot = createProbot(), webhooksPath } = {} as MiddlewareOptions,
 ): RequestListener {
   probot.load(appFn);
 
-  return createWebbhooksMiddleware(probot.webhooks, {
-    path: webhooksPath || "/",
+  return createWebhooksMiddleware(probot.webhooks, {
+    path: webhooksPath || probot.webhookPath || defaultWebhooksPath,
   });
 }
