@@ -1,16 +1,16 @@
 import type { Logger } from "pino";
-import {
+import type {
   WebhookError,
   EmitterWebhookEvent as WebhookEvent,
 } from "@octokit/webhooks";
 
 export function getErrorHandler(log: Logger) {
-  return (error: Error) => {
+  return (error: Error & { event?: WebhookEvent }) => {
     const errors = (
       error.name === "AggregateError" ? error : [error]
     ) as WebhookError[];
 
-    const event = (error as any).event as WebhookEvent;
+    const event = error.event;
 
     for (const error of errors) {
       const errMessage = (error.message || "").toLowerCase();
@@ -18,7 +18,7 @@ export function getErrorHandler(log: Logger) {
       if (errMessage.includes("x-hub-signature-256")) {
         log.error(
           error,
-          "Go to https://github.com/settings/apps/YOUR_APP and verify that the Webhook secret matches the value of the WEBHOOK_SECRET environment variable."
+          "Go to https://github.com/settings/apps/YOUR_APP and verify that the Webhook secret matches the value of the WEBHOOK_SECRET environment variable.",
         );
         continue;
       }
@@ -26,7 +26,7 @@ export function getErrorHandler(log: Logger) {
       if (errMessage.includes("pem") || errMessage.includes("json web token")) {
         log.error(
           error,
-          "Your private key (a .pem file or PRIVATE_KEY environment variable) or APP_ID is incorrect. Go to https://github.com/settings/apps/YOUR_APP, verify that APP_ID is set correctly, and generate a new PEM file if necessary."
+          "Your private key (a .pem file or PRIVATE_KEY environment variable) or APP_ID is incorrect. Go to https://github.com/settings/apps/YOUR_APP, verify that APP_ID is set correctly, and generate a new PEM file if necessary.",
         );
         continue;
       }
@@ -34,7 +34,7 @@ export function getErrorHandler(log: Logger) {
       log
         .child({
           name: "event",
-          id: event ? event.id : undefined,
+          id: event?.id,
         })
         .error(error);
     }
