@@ -14,8 +14,10 @@
  * app.log.fatal("Goodbye, cruel world!");
  * ```
  */
-import pino, { Logger, LoggerOptions } from "pino";
-import { getTransformStream, Options, LogLevel } from "@probot/pino";
+import { pino } from "pino";
+import type { Logger, LoggerOptions } from "pino";
+import { getTransformStream, type Options, type LogLevel } from "@probot/pino";
+import { rebindLog } from "./rebind-log.js";
 
 export type GetLogOptions = {
   level?: LogLevel;
@@ -31,9 +33,7 @@ export function getLog(options: GetLogOptions = {}): Logger {
     messageKey: logMessageKey || "msg",
   };
   const transform = getTransformStream(getTransformStreamOptions);
-  // @ts-ignore TODO: check out what's wrong here
   transform.pipe(pino.destination(1));
-  const log = pino(pinoOptions, transform);
 
-  return log;
+  return rebindLog(pino(pinoOptions, transform));
 }
