@@ -1,6 +1,6 @@
 import { Octokit } from "@octokit/core";
 import { enterpriseCompatibility } from "@octokit/plugin-enterprise-compatibility";
-import { RequestOptions } from "@octokit/types";
+import type { RequestOptions } from "@octokit/types";
 import { paginateRest } from "@octokit/plugin-paginate-rest";
 import { legacyRestEndpointMethods } from "@octokit/plugin-rest-endpoint-methods";
 import { retry } from "@octokit/plugin-retry";
@@ -8,29 +8,30 @@ import { throttling } from "@octokit/plugin-throttling";
 import { config } from "@probot/octokit-plugin-config";
 import { createProbotAuth } from "octokit-auth-probot";
 
-import { probotRequestLogging } from "./octokit-plugin-probot-request-logging";
-import { VERSION } from "../version";
+import { probotRequestLogging } from "./octokit-plugin-probot-request-logging.js";
+import { VERSION } from "../version.js";
 
 const defaultOptions = {
   authStrategy: createProbotAuth,
   throttle: {
-    onAbuseLimit: (
+    enabled: true,
+    onSecondaryRateLimit: (
       retryAfter: number,
       options: RequestOptions,
-      octokit: Octokit
+      octokit: Octokit,
     ) => {
       octokit.log.warn(
-        `Abuse limit hit with "${options.method} ${options.url}", retrying in ${retryAfter} seconds.`
+        `Secondary Rate limit hit with "${options.method} ${options.url}", retrying in ${retryAfter} seconds.`,
       );
       return true;
     },
     onRateLimit: (
       retryAfter: number,
       options: RequestOptions,
-      octokit: Octokit
+      octokit: Octokit,
     ) => {
       octokit.log.warn(
-        `Rate limit hit with "${options.method} ${options.url}", retrying in ${retryAfter} seconds.`
+        `Rate limit hit with "${options.method} ${options.url}", retrying in ${retryAfter} seconds.`,
       );
       return true;
     },
@@ -45,14 +46,18 @@ export const ProbotOctokit = Octokit.plugin(
   legacyRestEndpointMethods,
   enterpriseCompatibility,
   probotRequestLogging,
-  config
+  config,
 ).defaults((instanceOptions: any) => {
   // merge throttle options deeply
-  const options = Object.assign({}, defaultOptions, instanceOptions, {
-    throttle: instanceOptions.throttle
-      ? Object.assign({}, defaultOptions.throttle, instanceOptions.throttle)
-      : defaultOptions.throttle,
-  });
+  const options = {
+    ...defaultOptions,
+    ...instanceOptions,
+    ...{
+      throttle: { ...defaultOptions.throttle, ...instanceOptions?.throttle },
+    },
+  };
 
   return options;
 });
+
+export type ProbotOctokit = InstanceType<typeof ProbotOctokit>;
