@@ -1,28 +1,28 @@
-import LRUCache from "lru-cache";
-import { Logger } from "pino";
-import { EmitterWebhookEvent as WebhookEvent } from "@octokit/webhooks";
+import { LRUCache } from "lru-cache";
+import type { Logger } from "pino";
+import type { EmitterWebhookEvent as WebhookEvent } from "@octokit/webhooks";
 
-import { aliasLog } from "./helpers/alias-log";
-import { auth } from "./auth";
-import { getLog } from "./helpers/get-log";
-import { getProbotOctokitWithDefaults } from "./octokit/get-probot-octokit-with-defaults";
-import { getWebhooks } from "./octokit/get-webhooks";
-import { ProbotOctokit } from "./octokit/probot-octokit";
-import { VERSION } from "./version";
-import {
+import { auth } from "./auth.js";
+import { getLog } from "./helpers/get-log.js";
+import { getProbotOctokitWithDefaults } from "./octokit/get-probot-octokit-with-defaults.js";
+import { getWebhooks } from "./octokit/get-webhooks.js";
+import { ProbotOctokit } from "./octokit/probot-octokit.js";
+import { VERSION } from "./version.js";
+import type {
   ApplicationFunction,
   ApplicationFunctionOptions,
-  DeprecatedLogger,
   Options,
   ProbotWebhooks,
   State,
-} from "./types";
+} from "./types.js";
+import { defaultWebhooksPath } from "./server/server.js";
+import { rebindLog } from "./helpers/rebind-log.js";
 
-export type Constructor<T> = new (...args: any[]) => T;
+export type Constructor<T = any> = new (...args: any[]) => T;
 
 export class Probot {
   static version = VERSION;
-  static defaults<S extends Constructor<any>>(this: S, defaults: Options) {
+  static defaults<S extends Constructor>(this: S, defaults: Options) {
     const ProbotWithDefaults = class extends this {
       constructor(...args: any[]) {
         const options = args[0] || {};
@@ -34,15 +34,16 @@ export class Probot {
   }
 
   public webhooks: ProbotWebhooks;
-  public log: DeprecatedLogger;
+  public webhookPath: string;
+  public log: Logger;
   public version: String;
   public on: ProbotWebhooks["on"];
   public onAny: ProbotWebhooks["onAny"];
   public onError: ProbotWebhooks["onError"];
   public auth: (
     installationId?: number,
-    log?: Logger
-  ) => Promise<InstanceType<typeof ProbotOctokit>>;
+    log?: Logger,
+  ) => Promise<ProbotOctokit>;
 
   private state: State;
 
@@ -52,14 +53,16 @@ export class Probot {
     let level = options.logLevel;
     const logMessageKey = options.logMessageKey;
 
-    this.log = aliasLog(options.log || getLog({ level, logMessageKey }));
+    this.log = options.log
+      ? rebindLog(options.log)
+      : getLog({ level, logMessageKey });
 
     // TODO: support redis backend for access token cache if `options.redisConfig`
     const cache = new LRUCache<number, string>({
       // cache max. 15000 tokens, that will use less than 10mb memory
       max: 15000,
       // Cache for 1 minute less than GitHub expiry
-      maxAge: 1000 * 60 * 59,
+      ttl: 1000 * 60 * 59,
     });
 
     const Octokit = getProbotOctokitWithDefaults({
@@ -68,16 +71,20 @@ export class Probot {
       appId: Number(options.appId),
       privateKey: options.privateKey,
       cache,
-      log: this.log,
+      log: rebindLog(this.log),
       redisConfig: options.redisConfig,
       baseUrl: options.baseUrl,
     });
-    const octokit = new Octokit();
+    const octokitLogger = rebindLog(this.log.child({ name: "octokit" }));
+    const octokit = new Octokit({
+      request: options.request,
+      log: octokitLogger,
+    });
 
     this.state = {
       cache,
       githubToken: options.githubToken,
-      log: this.log,
+      log: rebindLog(this.log),
       Octokit,
       octokit,
       webhooks: {
@@ -87,11 +94,14 @@ export class Probot {
       privateKey: options.privateKey,
       host: options.host,
       port: options.port,
+      webhookPath: options.webhookPath || defaultWebhooksPath,
+      request: options.request,
     };
 
     this.auth = auth.bind(null, this.state);
 
     this.webhooks = getWebhooks(this.state);
+    this.webhookPath = this.state.webhookPath;
 
     this.on = this.webhooks.on;
     this.onAny = this.webhooks.onAny;
@@ -107,7 +117,7 @@ export class Probot {
 
   public async load(
     appFn: ApplicationFunction | ApplicationFunction[],
-    options: ApplicationFunctionOptions = {}
+    options: ApplicationFunctionOptions = {},
   ) {
     if (Array.isArray(appFn)) {
       for (const fn of appFn) {
