@@ -1,18 +1,21 @@
 import pkgConf from "pkg-conf";
 
-import { ApplicationFunction, Options, ServerOptions } from "./types";
-import { Probot } from "./index";
-import { setupAppFactory } from "./apps/setup";
-import { getLog, GetLogOptions } from "./helpers/get-log";
-import { readCliOptions } from "./bin/read-cli-options";
-import { readEnvOptions } from "./bin/read-env-options";
-import { Server } from "./server/server";
-import { defaultApp } from "./apps/default";
-import { resolveAppFunction } from "./helpers/resolve-app-function";
-import { isProduction } from "./helpers/is-production";
+import type { ApplicationFunction, Options, ServerOptions } from "./types.js";
+import { Logger, Probot, ProbotOctokit } from "./index.js";
+import { setupAppFactory } from "./apps/setup.js";
+import { getLog } from "./helpers/get-log.js";
+import { readCliOptions } from "./bin/read-cli-options.js";
+import { readEnvOptions } from "./bin/read-env-options.js";
+import { Server } from "./server/server.js";
+import { defaultApp } from "./apps/default.js";
+import { resolveAppFunction } from "./helpers/resolve-app-function.js";
+import { isProduction } from "./helpers/is-production.js";
+import { config as dotenvConfig } from "dotenv";
 
 type AdditionalOptions = {
-  env: Record<string, string | undefined>;
+  env?: NodeJS.ProcessEnv;
+  Octokit?: typeof ProbotOctokit;
+  log?: Logger;
 };
 
 /**
@@ -21,9 +24,9 @@ type AdditionalOptions = {
  */
 export async function run(
   appFnOrArgv: ApplicationFunction | string[],
-  additionalOptions?: AdditionalOptions
+  additionalOptions?: AdditionalOptions,
 ) {
-  require("dotenv").config();
+  dotenvConfig();
 
   const envOptions = readEnvOptions(additionalOptions?.env);
   const cliOptions = Array.isArray(appFnOrArgv)
@@ -55,15 +58,13 @@ export async function run(
     args,
   } = { ...envOptions, ...cliOptions };
 
-  const logOptions: GetLogOptions = {
+  const log = getLog({
     level,
     logFormat,
     logLevelInString,
     logMessageKey,
     sentryDsn,
-  };
-
-  const log = getLog(logOptions);
+  });
 
   const probotOptions: Options = {
     appId,
@@ -71,7 +72,8 @@ export async function run(
     redisConfig,
     secret,
     baseUrl,
-    log: log.child({ name: "probot" }),
+    log: additionalOptions?.log || log.child({ name: "probot" }),
+    Octokit: additionalOptions?.Octokit || undefined,
   };
 
   const serverOptions: ServerOptions = {
@@ -90,12 +92,12 @@ export async function run(
       if (!appId) {
         throw new Error(
           "App ID is missing, and is required to run in production mode. " +
-            "To resolve, ensure the APP_ID environment variable is set."
+            "To resolve, ensure the APP_ID environment variable is set.",
         );
       } else if (!privateKey) {
         throw new Error(
           "Certificate is missing, and is required to run in production mode. " +
-            "To resolve, ensure either the PRIVATE_KEY or PRIVATE_KEY_PATH environment variable is set and contains a valid certificate"
+            "To resolve, ensure either the PRIVATE_KEY or PRIVATE_KEY_PATH environment variable is set and contains a valid certificate",
         );
       }
     }
@@ -122,7 +124,7 @@ export async function run(
   if (Array.isArray(appFnOrArgv)) {
     const pkg = await pkgConf("probot");
 
-    const combinedApps: ApplicationFunction = async (app) => {
+    const combinedApps: ApplicationFunction = async (_app) => {
       await server.load(defaultApp);
 
       if (Array.isArray(pkg.apps)) {
