@@ -1,18 +1,21 @@
-import pinoHttp from "pino-http";
-import type { Logger } from "pino";
-import { v4 as uuidv4 } from "uuid";
+import { randomUUID as uuidv4 } from "node:crypto";
 
-export function getLoggingMiddleware(logger: Logger) {
+import { pinoHttp, startTime, type Options, type HttpLogger } from "pino-http";
+import type { Logger } from "pino";
+
+export function getLoggingMiddleware(
+  logger: Logger,
+  options?: Options,
+): HttpLogger {
   return pinoHttp({
+    ...options,
     logger: logger.child({ name: "http" }),
-    customSuccessMessage(res) {
-      const responseTime = Date.now() - res[pinoHttp.startTime];
-      // @ts-ignore
+    customSuccessMessage(_req, res) {
+      const responseTime = Date.now() - res[startTime];
       return `${res.req.method} ${res.req.url} ${res.statusCode} - ${responseTime}ms`;
     },
-    customErrorMessage(err, res) {
-      const responseTime = Date.now() - res[pinoHttp.startTime];
-      // @ts-ignore
+    customErrorMessage(_err, res) {
+      const responseTime = Date.now() - res[startTime];
       return `${res.req.method} ${res.req.url} ${res.statusCode} - ${responseTime}ms`;
     },
     genReqId: (req) =>
