@@ -1,11 +1,20 @@
-<!DOCTYPE html>
+export function importView({
+  name,
+  GH_HOST,
+  WEBHOOK_PROXY_URL = "",
+}: {
+  name?: string;
+  GH_HOST: string;
+  WEBHOOK_PROXY_URL?: string;
+}): string {
+  return `<!DOCTYPE html>
 <html lang="en" class="height-full" data-color-mode="auto" data-light-theme="light" data-dark-theme="dark">
 
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Import {{#if pkg.name }}{{ pkg.name }}{{else}}Your App{{/if}} | built with Probot</title>
+    <title>Import ${name || "Your App"} | built with Probot</title>
     <link rel="icon" href="/probot/static/probot-head.png">
     <link rel="stylesheet" href="/probot/static/primer.css">
 </head>
@@ -20,9 +29,9 @@
             <h3>Step 1:</h3>
             <p class="d-block mt-2">
                 Replace your app's Webhook URL with <br>
-                <b>{{ WEBHOOK_PROXY_URL }}</b>
+                <b>${WEBHOOK_PROXY_URL}</b>
             </p>
-            <a class="d-block mt-2" href="{{ GH_HOST }}/settings/apps" target="__blank" rel="noreferrer">
+            <a class="d-block mt-2" href="${GH_HOST}/settings/apps" target="__blank" rel="noreferrer">
                 You can do it here
             </a>
 
@@ -49,7 +58,7 @@
             <h4 class="alt-h4 text-gray-light">Need help?</h4>
             <div class="d-flex flex-justify-center mt-2">
                 <a href="https://probot.github.io/docs/" class="btn btn-outline mr-2">Documentation</a>
-                <a href="https://probot-slackin.herokuapp.com/" class="btn btn-outline">Chat on Slack</a>
+                <a href="https://github.com/probot/probot/discussions" class="btn btn-outline">Discuss on GitHub</a>
             </div>
         </div>
     </div>
@@ -81,4 +90,5 @@
     </script>
 </body>
 
-</html>
+</html>`;
+}
