@@ -1,37 +1,59 @@
-import Stream from "stream";
+import Stream from "node:stream";
 
-import {
+import type {
   EmitterWebhookEvent,
   EmitterWebhookEvent as WebhookEvent,
 } from "@octokit/webhooks";
 import Bottleneck from "bottleneck";
-import nock from "nock";
-import pino from "pino";
+import fetchMock from "fetch-mock";
+import { pino, type LogFn } from "pino";
+import { describe, expect, test, beforeEach, it, vi, type Mock } from "vitest";
 
-import { Probot, ProbotOctokit, Context } from "../src";
+import { Probot, ProbotOctokit, Context } from "../src/index.js";
 
-import webhookExamples from "@octokit/webhooks-examples";
-import { EmitterWebhookEventName } from "@octokit/webhooks/dist-types/types";
+import webhookExamples, {
+  type WebhookDefinition,
+} from "@octokit/webhooks-examples";
+import type { EmitterWebhookEventName } from "@octokit/webhooks/dist-types/types.js";
 
 const appId = 1;
 const privateKey = `-----BEGIN RSA PRIVATE KEY-----
-MIIBOQIBAAJBAIILhiN9IFpaE0pUXsesuuoaj6eeDiAqCiE49WB1tMB8ZMhC37kY
-Fl52NUYbUxb7JEf6pH5H9vqw1Wp69u78XeUCAwEAAQJAb88urnaXiXdmnIK71tuo
-/TyHBKt9I6Rhfzz0o9Gv7coL7a537FVDvV5UCARXHJMF41tKwj+zlt9EEUw7a1HY
-wQIhAL4F/VHWSPHeTgXYf4EaX2OlpSOk/n7lsFtL/6bWRzRVAiEArzJs2vopJitv
-A1yBjz3q2nX+zthk+GLXrJQkYOnIk1ECIHfeFV8TWm5gej1LxZquBTA5pINoqDVq
-NKZSuZEHqGEFAiB6EDrxkovq8SYGhIQsJeqkTMO8n94xhMRZlFmIQDokEQIgAq5U
-r1UQNnUExRh7ZT0kFbMfO9jKYZVlQdCL9Dn93vo=
+MIIEpAIBAAKCAQEA1c7+9z5Pad7OejecsQ0bu3aozN3tihPmljnnudb9G3HECdnH
+lWu2/a1gB9JW5TBQ+AVpum9Okx7KfqkfBKL9mcHgSL0yWMdjMfNOqNtrQqKlN4kE
+p6RD++7sGbzbfZ9arwrlD/HSDAWGdGGJTSOBM6pHehyLmSC3DJoR/CTu0vTGTWXQ
+rO64Z8tyXQPtVPb/YXrcUhbBp8i72b9Xky0fD6PkEebOy0Ip58XVAn2UPNlNOSPS
+ye+Qjtius0Md4Nie4+X8kwVI2Qjk3dSm0sw/720KJkdVDmrayeljtKBx6AtNQsSX
+gzQbeMmiqFFkwrG1+zx6E7H7jqIQ9B6bvWKXGwIDAQABAoIBAD8kBBPL6PPhAqUB
+K1r1/gycfDkUCQRP4DbZHt+458JlFHm8QL6VstKzkrp8mYDRhffY0WJnYJL98tr4
+4tohsDbqFGwmw2mIaHjl24LuWXyyP4xpAGDpl9IcusjXBxLQLp2m4AKXbWpzb0OL
+Ulrfc1ZooPck2uz7xlMIZOtLlOPjLz2DuejVe24JcwwHzrQWKOfA11R/9e50DVse
+hnSH/w46Q763y4I0E3BIoUMsolEKzh2ydAAyzkgabGQBUuamZotNfvJoDXeCi1LD
+8yNCWyTlYpJZJDDXooBU5EAsCvhN1sSRoaXWrlMSDB7r/E+aQyKua4KONqvmoJuC
+21vSKeECgYEA7yW6wBkVoNhgXnk8XSZv3W+Q0xtdVpidJeNGBWnczlZrummt4xw3
+xs6zV+rGUDy59yDkKwBKjMMa42Mni7T9Fx8+EKUuhVK3PVQyajoyQqFwT1GORJNz
+c/eYQ6VYOCSC8OyZmsBM2p+0D4FF2/abwSPMmy0NgyFLCUFVc3OECpkCgYEA5OAm
+I3wt5s+clg18qS7BKR2DuOFWrzNVcHYXhjx8vOSWV033Oy3yvdUBAhu9A1LUqpwy
+Ma+unIgxmvmUMQEdyHQMcgBsVs10dR/g2xGjMLcwj6kn+xr3JVIZnbRT50YuPhf+
+ns1ScdhP6upo9I0/sRsIuN96Gb65JJx94gQ4k9MCgYBO5V6gA2aMQvZAFLUicgzT
+u/vGea+oYv7tQfaW0J8E/6PYwwaX93Y7Q3QNXCoCzJX5fsNnoFf36mIThGHGiHY6
+y5bZPPWFDI3hUMa1Hu/35XS85kYOP6sGJjf4kTLyirEcNKJUWH7CXY+00cwvTkOC
+S4Iz64Aas8AilIhRZ1m3eQKBgQCUW1s9azQRxgeZGFrzC3R340LL530aCeta/6FW
+CQVOJ9nv84DLYohTVqvVowdNDTb+9Epw/JDxtDJ7Y0YU0cVtdxPOHcocJgdUGHrX
+ZcJjRIt8w8g/s4X6MhKasBYm9s3owALzCuJjGzUKcDHiO2DKu1xXAb0SzRcTzUCn
+7daCswKBgQDOYPZ2JGmhibqKjjLFm0qzpcQ6RPvPK1/7g0NInmjPMebP0K6eSPx0
+9/49J6WTD++EajN7FhktUSYxukdWaCocAQJTDNYP0K88G4rtC2IYy5JFn9SWz5oh
+x//0u+zd/R/QRUzLOw4N72/Hu+UG6MNt5iDZFCtapRaKt6OvSBwy8w==
 -----END RSA PRIVATE KEY-----`;
 
 const getPayloadExamples = <TName extends EmitterWebhookEventName>(
-  name: TName
+  name: TName,
 ) => {
-  return webhookExamples.filter((event) => event.name === name.split(".")[0])[0]
-    .examples as EmitterWebhookEvent<TName>["payload"][];
+  return (webhookExamples as unknown as WebhookDefinition[]).filter(
+    (event) => event.name === name.split(".")[0],
+  )[0].examples as EmitterWebhookEvent<TName>["payload"][];
 };
 const getPayloadExample = <TName extends EmitterWebhookEventName>(
-  name: TName
+  name: TName,
 ) => {
   const examples = getPayloadExamples<TName>(name);
   if (name.includes(".")) {
@@ -43,7 +65,6 @@ const getPayloadExample = <TName extends EmitterWebhookEventName>(
   }
   return examples[0];
 };
-// tslint:disable:no-empty
 describe("Probot", () => {
   let probot: Probot;
   let event: WebhookEvent<
@@ -52,7 +73,7 @@ describe("Probot", () => {
   let output: any;
 
   const streamLogsToOutput = new Stream.Writable({ objectMode: true });
-  streamLogsToOutput._write = (object, encoding, done) => {
+  streamLogsToOutput._write = (object, _encoding, done) => {
     output.push(JSON.parse(object));
     done();
   };
@@ -69,22 +90,26 @@ describe("Probot", () => {
 
   describe(".defaults()", () => {
     test("sets default options for constructor", async () => {
-      const mock = nock("https://api.github.com").get("/app").reply(200, {
-        id: 1,
+      const fetch = fetchMock.sandbox().getOnce("https://api.github.com/app", {
+        status: 200,
+        body: {
+          id: 1,
+        },
       });
 
       const MyProbot = Probot.defaults({ appId, privateKey });
-      const probot = new MyProbot();
+      const probot = new MyProbot({
+        request: { fetch },
+      });
       const octokit = await probot.auth();
       await octokit.apps.getAuthenticated();
-      expect(mock.activeMocks()).toStrictEqual([]);
     });
   });
 
   describe("constructor", () => {
     it("no options", () => {
       expect(() => new Probot()).toThrow(
-        "[@octokit/auth-app] appId option is required"
+        "[@octokit/auth-app] appId option is required",
       );
     });
 
@@ -101,8 +126,8 @@ describe("Probot", () => {
     it("shouldn't overwrite `options.throttle` passed to `{Octokit: ProbotOctokit.defaults(options)}`", () => {
       expect.assertions(1);
 
-      const MyOctokit = ProbotOctokit.plugin((octokit, options) => {
-        expect(options.throttle.enabled).toEqual(false);
+      const MyOctokit = ProbotOctokit.plugin((_octokit, options) => {
+        expect(options.throttle?.enabled).toEqual(true);
       }).defaults({
         appId,
         privateKey,
@@ -133,7 +158,7 @@ describe("Probot", () => {
     it("responds with the correct error if webhook secret does not match", async () => {
       expect.assertions(1);
 
-      probot.log.error = jest.fn();
+      probot.log.error = vi.fn() as LogFn;
       probot.webhooks.on("push", () => {
         throw new Error("X-Hub-Signature-256 does not match blob signature");
       });
@@ -141,16 +166,14 @@ describe("Probot", () => {
       try {
         await probot.webhooks.receive(event);
       } catch (e) {
-        expect(
-          (probot.log.error as jest.Mock).mock.calls[0][1]
-        ).toMatchSnapshot();
+        expect((probot.log.error as Mock).mock.calls[0][1]).toMatchSnapshot();
       }
     });
 
     it("responds with the correct error if webhook secret is not found", async () => {
       expect.assertions(1);
 
-      probot.log.error = jest.fn();
+      probot.log.error = vi.fn() as LogFn;
       probot.webhooks.on("push", () => {
         throw new Error("No X-Hub-Signature-256 found on request");
       });
@@ -158,66 +181,58 @@ describe("Probot", () => {
       try {
         await probot.webhooks.receive(event);
       } catch (e) {
-        expect(
-          (probot.log.error as jest.Mock).mock.calls[0][1]
-        ).toMatchSnapshot();
+        expect((probot.log.error as Mock).mock.calls[0][1]).toMatchSnapshot();
       }
     });
 
     it("responds with the correct error if webhook secret is wrong", async () => {
       expect.assertions(1);
 
-      probot.log.error = jest.fn();
+      probot.log.error = vi.fn() as LogFn;
       probot.webhooks.on("push", () => {
         throw Error(
-          "webhooks:receiver ignored: POST / due to missing headers: x-hub-signature-256"
+          "webhooks:receiver ignored: POST / due to missing headers: x-hub-signature-256",
         );
       });
 
       try {
         await probot.webhooks.receive(event);
       } catch (e) {
-        expect(
-          (probot.log.error as jest.Mock).mock.calls[0][1]
-        ).toMatchSnapshot();
+        expect((probot.log.error as Mock).mock.calls[0][1]).toMatchSnapshot();
       }
     });
 
     it("responds with the correct error if the PEM file is missing", async () => {
       expect.assertions(1);
 
-      probot.log.error = jest.fn();
+      probot.log.error = vi.fn() as LogFn;
       probot.webhooks.onAny(() => {
         throw new Error(
-          "error:0906D06C:PEM routines:PEM_read_bio:no start line"
+          "error:0906D06C:PEM routines:PEM_read_bio:no start line",
         );
       });
 
       try {
         await probot.webhooks.receive(event);
       } catch (e) {
-        expect(
-          (probot.log.error as jest.Mock).mock.calls[0][1]
-        ).toMatchSnapshot();
+        expect((probot.log.error as Mock).mock.calls[0][1]).toMatchSnapshot();
       }
     });
 
     it("responds with the correct error if the jwt could not be decoded", async () => {
       expect.assertions(1);
 
-      probot.log.error = jest.fn();
+      probot.log.error = vi.fn() as LogFn;
       probot.webhooks.onAny(() => {
         throw new Error(
-          '{"message":"A JSON web token could not be decoded","documentation_url":"https://developer.github.com/v3"}'
+          '{"message":"A JSON web token could not be decoded","documentation_url":"https://developer.github.com/v3"}',
         );
       });
 
       try {
         await probot.webhooks.receive(event);
       } catch (e) {
-        expect(
-          (probot.log.error as jest.Mock).mock.calls[0][1]
-        ).toMatchSnapshot();
+        expect((probot.log.error as Mock).mock.calls[0][1]).toMatchSnapshot();
       }
     });
   });
@@ -227,7 +242,7 @@ describe("Probot", () => {
       const appFn = async (app: Probot) => {
         const octokit = await app.auth();
         expect(octokit.request.endpoint.DEFAULTS.baseUrl).toEqual(
-          "https://notreallygithub.com/api/v3"
+          "https://notreallygithub.com/api/v3",
         );
       };
 
@@ -242,7 +257,7 @@ describe("Probot", () => {
       const appFn = async (app: Probot) => {
         const octokit = await app.auth();
         expect(octokit.request.endpoint.DEFAULTS.baseUrl).toEqual(
-          "https://notreallygithub.com/api/v3"
+          "https://notreallygithub.com/api/v3",
         );
       };
 
@@ -261,7 +276,7 @@ describe("Probot", () => {
       const appFn = async (app: Probot) => {
         const octokit = await app.auth();
         expect(octokit.request.endpoint.DEFAULTS.baseUrl).toEqual(
-          "http://notreallygithub.com/api/v3"
+          "http://notreallygithub.com/api/v3",
         );
       };
 
@@ -273,42 +288,48 @@ describe("Probot", () => {
     });
   });
 
-  describe.skip("options.redisConfig as string", () => {
-    it("sets throttle options", async () => {
-      expect.assertions(2);
+  describe.skipIf(process.env.REDIS_URL === undefined)(
+    "options.redisConfig as string",
+    () => {
+      it("sets throttle options", async () => {
+        expect.assertions(2);
 
-      probot = new Probot({
-        githubToken: "faketoken",
-        redisConfig: "test",
-        Octokit: ProbotOctokit.plugin((octokit, options) => {
-          expect(options.throttle.Bottleneck).toBe(Bottleneck);
-          expect(options.throttle.connection).toBeInstanceOf(
-            Bottleneck.IORedisConnection
-          );
-        }),
+        probot = new Probot({
+          githubToken: "faketoken",
+          redisConfig: process.env.REDIS_URL,
+          Octokit: ProbotOctokit.plugin((_octokit, options) => {
+            expect(options.throttle?.Bottleneck).toBe(Bottleneck);
+            expect(options.throttle?.connection).toBeInstanceOf(
+              Bottleneck.IORedisConnection,
+            );
+          }),
+        });
       });
-    });
-  });
+    },
+  );
 
-  describe.skip("redis configuration object", () => {
-    it("sets throttle options", async () => {
-      expect.assertions(2);
-      const redisConfig = {
-        host: "test",
-      };
+  describe.skipIf(process.env.REDIS_URL === undefined)(
+    "redis configuration object",
+    () => {
+      it("sets throttle options", async () => {
+        expect.assertions(2);
+        const redisConfig = {
+          host: process.env.REDIS_URL,
+        };
 
-      probot = new Probot({
-        githubToken: "faketoken",
-        redisConfig,
-        Octokit: ProbotOctokit.plugin((octokit, options) => {
-          expect(options.throttle.Bottleneck).toBe(Bottleneck);
-          expect(options.throttle.connection).toBeInstanceOf(
-            Bottleneck.IORedisConnection
-          );
-        }),
+        probot = new Probot({
+          githubToken: "faketoken",
+          redisConfig,
+          Octokit: ProbotOctokit.plugin((_octokit, options) => {
+            expect(options.throttle?.Bottleneck).toBe(Bottleneck);
+            expect(options.throttle?.connection).toBeInstanceOf(
+              Bottleneck.IORedisConnection,
+            );
+          }),
+        });
       });
-    });
-  });
+    },
+  );
 
   describe("on", () => {
     beforeEach(() => {
@@ -325,7 +346,7 @@ describe("Probot", () => {
         privateKey,
       });
 
-      const spy = jest.fn();
+      const spy = vi.fn();
       probot.on("pull_request", spy);
 
       expect(spy).toHaveBeenCalledTimes(0);
@@ -341,8 +362,14 @@ describe("Probot", () => {
         privateKey,
       });
 
-      const spy = jest.fn();
+      const spy = vi.fn();
       probot.on("pull_request.opened", spy);
+
+      const event: WebhookEvent<"pull_request.opened"> = {
+        id: "123-456",
+        name: "pull_request",
+        payload: getPayloadExample("pull_request.opened"),
+      };
 
       await probot.receive(event);
       expect(spy).toHaveBeenCalled();
@@ -354,7 +381,7 @@ describe("Probot", () => {
         privateKey,
       });
 
-      const spy = jest.fn();
+      const spy = vi.fn();
       probot.on("pull_request.closed", spy);
 
       await probot.receive(event);
@@ -367,7 +394,7 @@ describe("Probot", () => {
         privateKey,
       });
 
-      const spy = jest.fn();
+      const spy = vi.fn();
       probot.onAny(spy);
 
       await probot.receive(event);
@@ -380,13 +407,19 @@ describe("Probot", () => {
         privateKey,
       });
 
+      const event: WebhookEvent<"pull_request.opened"> = {
+        id: "123-456",
+        name: "pull_request",
+        payload: getPayloadExample("pull_request.opened"),
+      };
+
       const event2: WebhookEvent<"issues.opened"> = {
         id: "123",
         name: "issues",
         payload: getPayloadExample("issues.opened"),
       };
 
-      const spy = jest.fn();
+      const spy = vi.fn();
       probot.on(["pull_request.opened", "issues.opened"], spy);
 
       await probot.receive(event);
@@ -401,7 +434,7 @@ describe("Probot", () => {
         log: pino(streamLogsToOutput),
       });
 
-      const handler = jest.fn().mockImplementation((context) => {
+      const handler = vi.fn().mockImplementation((context) => {
         expect(context.log.info).toBeDefined();
         context.log.info("testing");
 
@@ -409,7 +442,7 @@ describe("Probot", () => {
           expect.objectContaining({
             id: context.id,
             msg: "testing",
-          })
+          }),
         );
       });
 
@@ -419,9 +452,40 @@ describe("Probot", () => {
     });
 
     it("returns an authenticated client for installation.created", async () => {
+      const fetch = fetchMock
+        .sandbox()
+        .postOnce("https://api.github.com/app/installations/1/access_tokens", {
+          status: 201,
+          body: {
+            token: "v1.1f699f1069f60xxx",
+            permissions: {
+              issues: "write",
+              contents: "read",
+            },
+          },
+        })
+        .getOnce(
+          function (url, opts) {
+            if (url === "https://api.github.com/") {
+              expect(
+                (opts.headers as Record<string, string>).authorization,
+              ).toEqual("token v1.1f699f1069f60xxx");
+              return true;
+            }
+            throw new Error("Should have matched");
+          },
+          {
+            status: 200,
+            body: {},
+          },
+        );
+
       const probot = new Probot({
         appId,
         privateKey,
+        request: {
+          fetch,
+        },
       });
 
       event = {
@@ -431,32 +495,35 @@ describe("Probot", () => {
       };
       event.payload.installation.id = 1;
 
-      const mock = nock("https://api.github.com")
-        .post("/app/installations/1/access_tokens")
-        .reply(201, {
-          token: "v1.1f699f1069f60xxx",
-          permissions: {
-            issues: "write",
-            contents: "read",
-          },
-        })
-        .get("/")
-        .matchHeader("authorization", "token v1.1f699f1069f60xxx")
-        .reply(200, {});
-
       probot.on("installation.created", async (context) => {
         await context.octokit.request("/");
       });
 
       await probot.receive(event);
-
-      expect(mock.activeMocks()).toStrictEqual([]);
     });
 
     it("returns an unauthenticated client for installation.deleted", async () => {
+      const fetch = fetchMock.sandbox().getOnce(
+        function (url, opts) {
+          if (url === "https://api.github.com/") {
+            expect(
+              (opts.headers as Record<string, string>).authorization,
+            ).toEqual(undefined);
+            return true;
+          }
+          throw new Error("Should have matched");
+        },
+        {
+          body: {},
+        },
+      );
+
       const probot = new Probot({
         appId,
         privateKey,
+        request: {
+          fetch,
+        },
       });
 
       event = {
@@ -466,46 +533,50 @@ describe("Probot", () => {
       };
       event.payload.installation.id = 1;
 
-      const mock = nock("https://api.github.com")
-        .get("/")
-        .matchHeader("authorization", (value) => value === undefined)
-        .reply(200, {});
-
       probot.on("installation.deleted", async (context) => {
         await context.octokit.request("/");
       });
 
       await probot.receive(event).catch(console.log);
-
-      expect(mock.activeMocks()).toStrictEqual([]);
     });
 
     it("returns an authenticated client for events without an installation", async () => {
+      const fetch = fetchMock.sandbox().mock(
+        function (url, opts) {
+          if (url === "https://api.github.com/") {
+            expect(
+              (opts.headers as Record<string, string>).authorization,
+            ).toEqual(undefined);
+            return true;
+          }
+          throw new Error("Should have matched");
+        },
+        {
+          body: {},
+        },
+      );
+
       const probot = new Probot({
         appId,
         privateKey,
+        request: {
+          fetch,
+        },
       });
 
       event = {
         id: "123-456",
         name: "check_run",
         payload: getPayloadExamples("check_run").filter(
-          (event) => typeof event.installation === "undefined"
+          (event) => typeof event.installation === "undefined",
         )[0],
       };
-
-      const mock = nock("https://api.github.com")
-        .get("/")
-        .matchHeader("authorization", (value) => value === undefined)
-        .reply(200, {});
 
       probot.on("check_run", async (context) => {
         await context.octokit.request("/");
       });
 
       await probot.receive(event).catch(console.log);
-
-      expect(mock.activeMocks()).toStrictEqual([]);
     });
   });
 
@@ -524,7 +595,7 @@ describe("Probot", () => {
         privateKey,
       });
 
-      const spy = jest.fn();
+      const spy = vi.fn();
       probot.on("pull_request", spy);
 
       await probot.receive(event);
@@ -538,7 +609,7 @@ describe("Probot", () => {
         privateKey,
       });
 
-      const spy = jest.fn();
+      const spy = vi.fn();
       probot.on("pull_request", () => {
         return new Promise((resolve) => {
           setTimeout(() => {
@@ -568,8 +639,24 @@ describe("Probot", () => {
         await probot.receive(event);
         throw new Error("expected error to be raised from app");
       } catch (error) {
-        expect(error.message).toMatch(/error from app/);
+        expect((error as Error).message).toMatch(/error from app/);
       }
+    });
+
+    it("passes logger to webhooks", async () => {
+      const probot = new Probot({
+        appId,
+        privateKey,
+        log: pino(streamLogsToOutput),
+      });
+
+      // @ts-expect-error
+      probot.on("unknown-event", () => {});
+
+      expect(output.length).toEqual(1);
+      expect(output[0].msg).toEqual(
+        '"unknown-event" is not a known webhook name (https://developer.github.com/v3/activity/events/types/)',
+      );
     });
   });
 });
