@@ -1,23 +1,28 @@
-import { resolveAppFunction } from "../src/helpers/resolve-app-function";
+import { resolveAppFunction } from "../src/helpers/resolve-app-function.js";
+import { describe, expect, vi, it } from "vitest";
 
-const stubAppFnPath = require.resolve("./fixtures/plugin/stub-plugin");
+const stubAppFnPath = require.resolve("./fixtures/plugin/stub-plugin.ts");
 const stubTranspiledAppFnPath = require.resolve(
-  "./fixtures/plugin/stub-typescript-transpiled-plugin"
+  "./fixtures/plugin/stub-typescript-transpiled-plugin.ts",
 );
 const basedir = process.cwd();
 
 describe("resolver", () => {
   it("loads the module at the resolved path", async () => {
-    const stubResolver = jest.fn().mockReturnValue(stubAppFnPath);
-    const module = await resolveAppFunction("foo", { resolver: stubResolver });
-    expect(module).toBe(require(stubAppFnPath));
-    expect(stubResolver).toHaveBeenCalledWith("foo", { basedir });
+    const stubResolver = vi.fn().mockReturnValue(stubAppFnPath);
+    const module = await resolveAppFunction("foo", {
+      resolver: stubResolver as unknown as RequireResolve,
+    });
+    expect(module).toBeInstanceOf(Function);
+    expect(stubResolver).toHaveBeenCalledWith("foo", { paths: [basedir] });
   });
 
   it("loads module transpiled from TypeScript (https://github.com/probot/probot/issues/1447)", async () => {
-    const stubResolver = jest.fn().mockReturnValue(stubTranspiledAppFnPath);
-    const module = await resolveAppFunction("foo", { resolver: stubResolver });
-    expect(module).toBe(require(stubTranspiledAppFnPath).default);
-    expect(stubResolver).toHaveBeenCalledWith("foo", { basedir });
+    const stubResolver = vi.fn().mockReturnValue(stubTranspiledAppFnPath);
+    const module = await resolveAppFunction("foo", {
+      resolver: stubResolver as unknown as RequireResolve,
+    });
+    expect(module).toBeInstanceOf(Function);
+    expect(stubResolver).toHaveBeenCalledWith("foo", { paths: [basedir] });
   });
 });
