@@ -1,28 +1,28 @@
-import express, { Response } from "express";
-// tslint:disable-next-line:no-var-requires
+import { randomInt } from "node:crypto";
+import http from "node:http";
+import net from "node:net";
+
+import express, { type Response } from "express";
 const sse: (
   req: express.Request,
   res: express.Response,
-  next: express.NextFunction
+  next: express.NextFunction,
 ) => void = require("connect-sse")();
-import EventSource from "eventsource";
-import http from "http";
-import net from "net";
-import nock from "nock";
-import { getLog } from "../src/helpers/get-log";
-import { createWebhookProxy } from "../src/helpers/webhook-proxy";
+import fetchMock from "fetch-mock";
+import { describe, expect, afterEach, test, vi } from "vitest";
+import { getLog } from "../src/helpers/get-log.js";
+import { createWebhookProxy } from "../src/helpers/webhook-proxy.js";
 
-const targetPort = 999999;
+let targetPort = 999999;
 
 interface SSEResponse extends Response {
   json(body: any, status?: string): this;
 }
 
-jest.setTimeout(10000);
-
 describe("webhook-proxy", () => {
-  // tslint:disable-next-line:one-variable-per-declaration
-  let emit: SSEResponse["json"], proxy: EventSource, server: http.Server;
+  let emit: SSEResponse["json"];
+  let proxy: EventSource;
+  let server: http.Server;
 
   afterEach(() => {
     server && server.close();
@@ -30,59 +30,111 @@ describe("webhook-proxy", () => {
   });
 
   describe("with a valid proxy server", () => {
-    beforeEach((done) => {
+    test("forwards events to server", async () => {
+      let readyPromise = {
+        promise: undefined,
+        reject: undefined,
+        resolve: undefined,
+      } as {
+        promise?: Promise<any>;
+        resolve?: (value?: any) => any;
+        reject?: (reason?: any) => any;
+      };
+
+      readyPromise.promise = new Promise((resolve, reject) => {
+        readyPromise.resolve = resolve;
+        readyPromise.reject = reject;
+      });
+
+      let finishedPromise = {
+        promise: undefined,
+        reject: undefined,
+        resolve: undefined,
+      } as {
+        promise?: Promise<any>;
+        resolve?: (value?: any) => any;
+        reject?: (reason?: any) => any;
+      };
+
+      finishedPromise.promise = new Promise((resolve, reject) => {
+        finishedPromise.resolve = resolve;
+        finishedPromise.reject = reject;
+      });
+
       const app = express();
 
-      app.get("/events", sse, (req, res: SSEResponse) => {
+      app.get("/events", sse, (_req, res: SSEResponse) => {
         res.json({}, "ready");
         emit = res.json;
       });
 
-      server = app.listen(0, () => {
-        const url = `http://127.0.0.1:${
-          (server.address() as net.AddressInfo).port
-        }/events`;
-        proxy = createWebhookProxy({
+      server = app.listen(0, async () => {
+        targetPort = (server.address() as net.AddressInfo).port;
+        const url = `http://127.0.0.1:${targetPort}/events`;
+
+        const fetch = fetchMock
+          .sandbox()
+          .postOnce(`http://localhost:${targetPort}/test`, {
+            status: 200,
+            then: () => {
+              finishedPromise.resolve!();
+            },
+          });
+
+        proxy = (await createWebhookProxy({
           url,
           port: targetPort,
           path: "/test",
           logger: getLog({ level: "fatal" }),
-        })!;
+          fetch,
+        })) as EventSource;
 
         // Wait for proxy to be ready
-        proxy.addEventListener("ready", () => done());
+        proxy.addEventListener("ready", readyPromise.resolve!);
       });
-    });
 
-    test("forwards events to server", (done) => {
-      nock(`http://localhost:${targetPort}`)
-        .post("/test")
-        .reply(200, () => {
-          done();
-        });
-
-      const body = { action: "foo" };
+      await readyPromise.promise;
 
       emit({
-        body,
+        body: { action: "foo" },
         "x-github-event": "test",
       });
+
+      await finishedPromise.promise;
     });
   });
 
-  test("logs an error when the proxy server is not found", (done) => {
-    const url = "http://bad.proxy/events";
-    nock("http://bad.proxy").get("/events").reply(404);
+  test("logs an error when the proxy server is not found", async () => {
+    expect.assertions(2);
 
-    const log = getLog({ level: "fatal" }).child({});
-    log.error = jest.fn();
+    let finishedPromise = {
+      promise: undefined,
+      reject: undefined,
+      resolve: undefined,
+    } as {
+      promise?: Promise<any>;
+      resolve?: (value?: any) => any;
+      reject?: (reason?: any) => any;
+    };
 
-    proxy = createWebhookProxy({ url, logger: log })!;
-
-    proxy.addEventListener("error", (error: any) => {
-      expect(error.status).toBe(404);
-      expect(log.error).toHaveBeenCalledWith(error);
-      done();
+    finishedPromise.promise = new Promise((resolve, reject) => {
+      finishedPromise.resolve = resolve;
+      finishedPromise.reject = reject;
     });
+
+    const url = `http://bad.n${randomInt(1e10).toString(36)}.proxy/events`;
+
+    const logger = getLog({ level: "fatal" }).child({});
+    logger.error = vi.fn() as any;
+
+    createWebhookProxy({ url, logger })!.then((proxy) => {
+      (proxy as EventSource).addEventListener("error", (error: any) => {
+        expect(error.message).toMatch(/^getaddrinfo ENOTFOUND/);
+        expect(logger.error).toHaveBeenCalledWith(error);
+        finishedPromise.resolve!();
+      });
+    });
+
+    await finishedPromise.promise;
   });
 });
